Allow alias and link to handle regex replacements
Allows for customizable regex routing / redirects
Example:

| Alias      | Link                                                   |
| ---------- | ------------------------------------------------------ |
| pr/(\d+)   | https://github.com/mongodb-labs/buildfest-2022/pull/$1 |
| JAVA-(\d+) | https://jira.mongodb.org/browse/JAVA-$1                |

Regex's come with a cost as cannot use `findOneAndUpdate` so there is
an accountancy cost.

Currently it routes using:

A `findOneAndUpdate` for simple urls.
If fails tries to fina a URL via an aggregation using `$regexFind`.
If successful then does an updateOne to increment the counter.